requirements.txt
aiohttp==3.6.2
aiohttp-json-rpc==0.12.2
aiosqlite==0.11.0
appdirs==1.4.3
apsw-wheels==3.30.1.post3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
contextlib2==0.5.5
discord.py==1.3.1
distro==1.4.0
fuzzywuzzy==0.17.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.3
pycparser==2.19
pylint==2.4.4
PyNaCl==1.3.0
python-Levenshtein-wheels==0.13.1
pytz==2019.3
PyYAML==5.3
Red-DiscordBot==3.3.0
Red-Lavalink==0.4.2
schema==0.7.1
six==1.14.0
tqdm==4.41.1
uvloop==0.14.0
websockets==6.0
wrapt==1.11.2
yarl==1.4.2
